(Phoenix) Update dependencies

+++ jmix-gridexport/gridexport-ui/src/main/java/io/jmix/gridexportui/exporter/excel/ExcelAutoColumnSizer.java
@@ -42,13 +42,13 @@ public class ExcelAutoColumnSizer {
     private Graphics2D graphics = null;
 
     /** Maps a Short (HSSF font index) to a FontMetrics object */
-    private Map<Short, FontMetrics> fontMetrics = new HashMap<>();
+    private Map<Integer, FontMetrics> fontMetrics = new HashMap<>();
 
     private short currentWidth = WIDTH_MIN;
 
     private FontMetrics getFontMetrics(Font hf){
         FontMetrics fm;
-        Short pFont = hf.getIndex();
+        int pFont = hf.getIndex();
 
         fm = fontMetrics.get(pFont);
         if (fm == null) {
